+++ CONTRIBUTORS.txt
@@ -1,5 +1,5 @@
 For the purpose of tracking copyright, this is the list of individuals and
-organizations who have contributed source code to SwiftAwsLambda.
+organizations who have contributed source code to SwiftAWSLambdaRuntime.
 
 For employees of an organization/company where the copyright of work done
 by employees of that company is held by the company itself, only the company
