+++ Tests/AWSLambdaRuntimeTests/Lambda+CodeableTest.swift
@@ -1,33 +1,33 @@
 //===----------------------------------------------------------------------===//
 //
-// This source file is part of the SwiftAwsLambda open source project
+// This source file is part of the SwiftAWSLambdaRuntime open source project
 //
-// Copyright (c) 2017-2018 Apple Inc. and the SwiftAwsLambda project authors
+// Copyright (c) 2017-2018 Apple Inc. and the SwiftAWSLambdaRuntime project authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of SwiftAwsLambda project authors
+// See CONTRIBUTORS.txt for the list of SwiftAWSLambdaRuntime project authors
 //
 // SPDX-License-Identifier: Apache-2.0
 //
 //===----------------------------------------------------------------------===//
 
+@testable import AWSLambdaRuntime
 import NIO
-@testable import SwiftAwsLambda
 import XCTest
 
-class StringLambdaTest: XCTestCase {
+class CodableLambdaTest: XCTestCase {
     func testCallbackSuccess() {
         let server = MockLambdaServer(behavior: Behavior())
         XCTAssertNoThrow(try server.start().wait())
         defer { XCTAssertNoThrow(try server.stop().wait()) }
 
         struct Handler: LambdaHandler {
-            typealias In = String
-            typealias Out = String
+            typealias In = Request
+            typealias Out = Response
 
-            func handle(context: Lambda.Context, payload: String, callback: (Result<String, Error>) -> Void) {
-                callback(.success(payload))
+            func handle(context: Lambda.Context, payload: Request, callback: (Result<Response, Error>) -> Void) {
+                callback(.success(Response(requestId: payload.requestId)))
             }
         }
 
@@ -43,10 +43,10 @@ class StringLambdaTest: XCTestCase {
         defer { XCTAssertNoThrow(try server.stop().wait()) }
 
         struct Handler: LambdaHandler {
-            typealias In = String
+            typealias In = Request
             typealias Out = Void
 
-            func handle(context: Lambda.Context, payload: String, callback: (Result<Void, Error>) -> Void) {
+            func handle(context: Lambda.Context, payload: Request, callback: (Result<Void, Error>) -> Void) {
                 callback(.success(()))
             }
         }
@@ -63,10 +63,10 @@ class StringLambdaTest: XCTestCase {
         defer { XCTAssertNoThrow(try server.stop().wait()) }
 
         struct Handler: LambdaHandler {
-            typealias In = String
-            typealias Out = String
+            typealias In = Request
+            typealias Out = Response
 
-            func handle(context: Lambda.Context, payload: String, callback: (Result<String, Error>) -> Void) {
+            func handle(context: Lambda.Context, payload: Request, callback: (Result<Response, Error>) -> Void) {
                 callback(.failure(TestError("boom")))
             }
         }
@@ -83,11 +83,11 @@ class StringLambdaTest: XCTestCase {
         defer { XCTAssertNoThrow(try server.stop().wait()) }
 
         struct Handler: EventLoopLambdaHandler {
-            typealias In = String
-            typealias Out = String
+            typealias In = Request
+            typealias Out = Response
 
-            func handle(context: Lambda.Context, payload: String) -> EventLoopFuture<String> {
-                context.eventLoop.makeSucceededFuture(payload)
+            func handle(context: Lambda.Context, payload: Request) -> EventLoopFuture<Response> {
+                context.eventLoop.makeSucceededFuture(Response(requestId: payload.requestId))
             }
         }
 
@@ -103,10 +103,10 @@ class StringLambdaTest: XCTestCase {
         defer { XCTAssertNoThrow(try server.stop().wait()) }
 
         struct Handler: EventLoopLambdaHandler {
-            typealias In = String
+            typealias In = Request
             typealias Out = Void
 
-            func handle(context: Lambda.Context, payload: String) -> EventLoopFuture<Void> {
+            func handle(context: Lambda.Context, payload: Request) -> EventLoopFuture<Void> {
                 context.eventLoop.makeSucceededFuture(())
             }
         }
@@ -123,10 +123,10 @@ class StringLambdaTest: XCTestCase {
         defer { XCTAssertNoThrow(try server.stop().wait()) }
 
         struct Handler: EventLoopLambdaHandler {
-            typealias In = String
-            typealias Out = String
+            typealias In = Request
+            typealias Out = Response
 
-            func handle(context: Lambda.Context, payload: String) -> EventLoopFuture<String> {
+            func handle(context: Lambda.Context, payload: Request) -> EventLoopFuture<Response> {
                 context.eventLoop.makeFailedFuture(TestError("boom"))
             }
         }
@@ -144,8 +144,8 @@ class StringLambdaTest: XCTestCase {
 
         let maxTimes = Int.random(in: 1 ... 10)
         let configuration = Lambda.Configuration(lifecycle: .init(maxTimes: maxTimes))
-        let result = Lambda.run(configuration: configuration) { (_, payload: String, callback) in
-            callback(.success(payload))
+        let result = Lambda.run(configuration: configuration) { (_, payload: Request, callback) in
+            callback(.success(Response(requestId: payload.requestId)))
         }
         assertLambdaLifecycleResult(result, shoudHaveRun: maxTimes)
     }
@@ -157,7 +157,7 @@ class StringLambdaTest: XCTestCase {
 
         let maxTimes = Int.random(in: 1 ... 10)
         let configuration = Lambda.Configuration(lifecycle: .init(maxTimes: maxTimes))
-        let result = Lambda.run(configuration: configuration) { (_, _: String, callback) in
+        let result = Lambda.run(configuration: configuration) { (_, _: Request, callback: (Result<Void, Error>) -> Void) in
             callback(.success(()))
         }
         assertLambdaLifecycleResult(result, shoudHaveRun: maxTimes)
@@ -170,7 +170,7 @@ class StringLambdaTest: XCTestCase {
 
         let maxTimes = Int.random(in: 1 ... 10)
         let configuration = Lambda.Configuration(lifecycle: .init(maxTimes: maxTimes))
-        let result: Result<Int, Error> = Lambda.run(configuration: configuration) { (_, _: String, callback: (Result<String, Error>) -> Void) in
+        let result: Result<Int, Error> = Lambda.run(configuration: configuration) { (_, _: Request, callback: (Result<Response, Error>) -> Void) in
             callback(.failure(TestError("boom")))
         }
         assertLambdaLifecycleResult(result, shoudHaveRun: maxTimes)
@@ -182,14 +182,14 @@ class StringLambdaTest: XCTestCase {
         defer { XCTAssertNoThrow(try server.stop().wait()) }
 
         struct Handler: LambdaHandler {
-            typealias In = String
-            typealias Out = String
+            typealias In = Request
+            typealias Out = Response
 
             init(eventLoop: EventLoop) throws {
                 throw TestError("kaboom")
             }
 
-            func handle(context: Lambda.Context, payload: String, callback: (Result<String, Error>) -> Void) {
+            func handle(context: Lambda.Context, payload: Request, callback: (Result<Response, Error>) -> Void) {
                 callback(.failure(TestError("should not be called")))
             }
         }
@@ -199,6 +199,7 @@ class StringLambdaTest: XCTestCase {
     }
 }
 
+// TODO: taking advantage of the fact we know the serialization is json
 private struct Behavior: LambdaServerBehavior {
     let requestId: String
     let payload: String
@@ -211,18 +212,38 @@ private struct Behavior: LambdaServerBehavior {
     }
 
     func getWork() -> GetWorkResult {
-        .success((requestId: self.requestId, payload: self.payload))
+        guard let payload = try? JSONEncoder().encode(Request(requestId: requestId)) else {
+            XCTFail("encoding error")
+            return .failure(.internalServerError)
+        }
+        guard let payloadAsString = String(data: payload, encoding: .utf8) else {
+            XCTFail("encoding error")
+            return .failure(.internalServerError)
+        }
+        return .success((requestId: self.requestId, payload: payloadAsString))
     }
 
     func processResponse(requestId: String, response: String?) -> Result<Void, ProcessResponseError> {
-        XCTAssertEqual(self.requestId, requestId, "expecting requestId to match")
         switch self.result {
-        case .success(let expected):
-            XCTAssertEqual(expected, response, "expecting response to match")
+        case .success(let expected) where expected != nil:
+            guard let data = response?.data(using: .utf8) else {
+                XCTFail("decoding error")
+                return .failure(.internalServerError)
+            }
+            guard let response = try? JSONDecoder().decode(Response.self, from: data) else {
+                XCTFail("decoding error")
+                return .failure(.internalServerError)
+            }
+            XCTAssertEqual(self.requestId, response.requestId, "expecting requestId to match")
+            return .success(())
+        case .success(let expected) where expected == nil:
+            XCTAssertNil(response)
             return .success(())
         case .failure:
             XCTFail("unexpected to fail, but succeeded with: \(response ?? "undefined")")
             return .failure(.internalServerError)
+        default:
+            preconditionFailure("invalid state")
         }
     }
 
@@ -241,5 +262,19 @@ private struct Behavior: LambdaServerBehavior {
     func processInitError(error: ErrorResponse) -> Result<Void, ProcessErrorError> {
         XCTFail("should not report init error")
         return .failure(.internalServerError)
+    }
+}
+
+private struct Request: Codable {
+    let requestId: String
+    init(requestId: String) {
+        self.requestId = requestId
+    }
+}
+
+private struct Response: Codable {
+    let requestId: String
+    init(requestId: String) {
+        self.requestId = requestId
     }
 }
