+++ Tests/AWSLambdaRuntimeTests/MockLambdaServer.swift
@@ -1,22 +1,22 @@
 //===----------------------------------------------------------------------===//
 //
-// This source file is part of the SwiftAwsLambda open source project
+// This source file is part of the SwiftAWSLambdaRuntime open source project
 //
-// Copyright (c) 2017-2018 Apple Inc. and the SwiftAwsLambda project authors
+// Copyright (c) 2017-2018 Apple Inc. and the SwiftAWSLambdaRuntime project authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of SwiftAwsLambda project authors
+// See CONTRIBUTORS.txt for the list of SwiftAWSLambdaRuntime project authors
 //
 // SPDX-License-Identifier: Apache-2.0
 //
 //===----------------------------------------------------------------------===//
 
+@testable import AWSLambdaRuntime
 import Foundation // for JSON
 import Logging
 import NIO
 import NIOHTTP1
-@testable import SwiftAwsLambda
 
 internal final class MockLambdaServer {
     private let logger = Logger(label: "MockLambdaServer")
